requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.0
chromadb==0.4.22
beautifulsoup4==4.12.2
